+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 - Chg #208: Rename `getAccessTree()` to `getHierarchy()` in `ItemsStorageInterface` (@arogachev)
 - Enh #252: Return `$this` instead of throwing "already assigned" exception in `Manager::assign()` (@arogachev)
 - Enh #248: Add `SimpleRuleFactory` (@arogachev)
+- Enh #251: Allow checking for user's roles in `ManagerInterface::userHasPermission()` (@arogachev)
 
 ## 1.0.2 April 20, 2023